Refactor lobby service to import types and update game metadata structure
