+++ src/types/index.ts
@@ -3,4 +3,21 @@ export type User = {
     name: string;
 };
 
+export type GameTypeMetadata = {
+    type: string;
+    displayName: string;
+    maxPlayers: number;
+    minPlayers: number;
+    requiresTeams: boolean;
+    numTeams?: number;
+    playersPerTeam?: number;
+    description?: string;
+};
+
+export interface CreateAndJoinRoomResponse {
+    roomId: string;
+    userId: string;
+    roomCode: string;
+}
+
 export * from "./games";
